requirements.txt
google-generativeai>=0.5.0,<0.7
PyYAML>=6.0
matplotlib>=3.7
seaborn>=0.12
pandas>=2.0
Markdown>=3.5
WeasyPrint>=61.0
Pygments>=2.16
